requirements.txt
opencv-python
scenedetect[opencv]
openai-whisper
open-clip-torch
librosa
Pillow
numpy
pyyaml
openai
